Added features
Cards class :- function to get the card number which will return numbers instead of Letter.
Command class :-  Added a function for executing multiple transaction. Also added the ability to place cards if the play deck is empty.
PlayArea class :- added the ability to return card number also for alphabets. Also added the function to return the status of the card, only the first card in the deck by peek() method